Fix: proxy directory
